test to find gas cost to solve bounty
